Finition inscription et redirect et mise a jour home
finition d'inscription, redirect et mise a jour de home
