+++ src/main/resources/templates/Connection.html
@@ -26,5 +26,10 @@
 	<div>
 		<a href="/inscription">Inscription</a>
 	</div>
+	<br>
+	<br>
+	<div>
+		<a href="/home">Retour</a>
+	</div>
 </body>
 </html>
