+++ src/main/resources/templates/Inscription.html
@@ -42,5 +42,10 @@
 	<div>
 		<a href="/connection">Connection</a>
 	</div>
+	<br>
+	<br>
+	<div>
+		<a href="/home">Retour</a>
+	</div>
 </body>
 </html>
